LayoutTests/transitions/transitions-parsing-expected.txt

Test the parsing and the computed style values of the transitions properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Valid transition-property values.
PASS computedStyle.transitionProperty is 'all'
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is 'none'
PASS computedStyle.transitionProperty is 'none'
PASS style.webkitTransitionProperty is 'none'
PASS computedStyle.webkitTransitionProperty is 'none'
PASS style.transitionProperty is 'all'
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is 'all'
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is 'background-position'
PASS computedStyle.transitionProperty is 'background-position'
PASS style.webkitTransitionProperty is 'background-position'
PASS computedStyle.webkitTransitionProperty is 'background-position'
PASS style.transitionProperty is 'background-position, font-size'
PASS computedStyle.transitionProperty is 'background-position, font-size'
PASS style.webkitTransitionProperty is 'background-position, font-size'
PASS computedStyle.webkitTransitionProperty is 'background-position, font-size'
PASS style.transitionProperty is 'background-position, font-size, color'
PASS computedStyle.transitionProperty is 'background-position, font-size, color'
PASS style.webkitTransitionProperty is 'background-position, font-size, color'
PASS computedStyle.webkitTransitionProperty is 'background-position, font-size, color'
PASS style.transitionProperty is 'all, font-size, color'
PASS computedStyle.transitionProperty is 'all, font-size, color'
PASS style.webkitTransitionProperty is 'all, font-size, color'
PASS computedStyle.webkitTransitionProperty is 'all, font-size, color'
PASS style.transitionProperty is 'font-size, color, all'
PASS computedStyle.transitionProperty is 'font-size, color, all'
PASS style.webkitTransitionProperty is 'font-size, color, all'
PASS computedStyle.webkitTransitionProperty is 'font-size, color, all'
PASS style.transitionProperty is 'font-size, all, color'
PASS computedStyle.transitionProperty is 'font-size, all, color'
PASS style.webkitTransitionProperty is 'font-size, all, color'
PASS computedStyle.webkitTransitionProperty is 'font-size, all, color'
Invalid transition-property values.
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
PASS style.transitionProperty is ''
PASS computedStyle.transitionProperty is 'all'
PASS style.webkitTransitionProperty is ''
PASS computedStyle.webkitTransitionProperty is 'all'
Valid transition-duration values.
PASS computedStyle.transitionDuration is '0s'
PASS computedStyle.webkitTransitionDuration is '0s'
PASS style.transitionDuration is '0s'
PASS computedStyle.transitionDuration is '0s'
PASS style.webkitTransitionDuration is '0s'
PASS computedStyle.webkitTransitionDuration is '0s'
PASS style.transitionDuration is '5s'
PASS computedStyle.transitionDuration is '5s'
PASS style.webkitTransitionDuration is '5s'
PASS computedStyle.webkitTransitionDuration is '5s'
PASS style.transitionDuration is '10ms'
PASS computedStyle.transitionDuration is '0.01s'
PASS style.webkitTransitionDuration is '10ms'
PASS computedStyle.webkitTransitionDuration is '0.01s'
PASS style.transitionDuration is '10ms, 20s'
PASS computedStyle.transitionDuration is '0.01s, 20s'
PASS style.webkitTransitionDuration is '10ms, 20s'
PASS computedStyle.webkitTransitionDuration is '0.01s, 20s'
Invalid transition-duration values.
PASS style.transitionDuration is ''
PASS computedStyle.transitionDuration is '0s'
PASS style.webkitTransitionDuration is ''
PASS computedStyle.webkitTransitionDuration is '0s'
PASS style.transitionDuration is ''
PASS computedStyle.transitionDuration is '0s'
PASS style.webkitTransitionDuration is ''
PASS computedStyle.webkitTransitionDuration is '0s'
PASS style.transitionDuration is ''
PASS computedStyle.transitionDuration is '0s'
PASS style.webkitTransitionDuration is ''
PASS computedStyle.webkitTransitionDuration is '0s'
PASS style.transitionDuration is ''
PASS computedStyle.transitionDuration is '0s'
PASS style.webkitTransitionDuration is ''
PASS computedStyle.webkitTransitionDuration is '0s'
PASS style.transitionDuration is ''
PASS computedStyle.transitionDuration is '0s'
PASS style.webkitTransitionDuration is ''
PASS computedStyle.webkitTransitionDuration is '0s'
PASS style.transitionDuration is ''
PASS computedStyle.transitionDuration is '0s'
PASS style.webkitTransitionDuration is ''
PASS computedStyle.webkitTransitionDuration is '0s'
PASS style.transitionDuration is ''
PASS computedStyle.transitionDuration is '0s'
PASS style.webkitTransitionDuration is ''
PASS computedStyle.webkitTransitionDuration is '0s'
Valid transition-timing-function values.
PASS computedStyle.transitionTimingFunction is 'ease'
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is 'linear'
PASS computedStyle.transitionTimingFunction is 'linear'
PASS style.webkitTransitionTimingFunction is 'linear'
PASS computedStyle.webkitTransitionTimingFunction is 'linear'
PASS style.transitionTimingFunction is 'ease'
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is 'ease'
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is 'ease-in'
PASS computedStyle.transitionTimingFunction is 'ease-in'
PASS style.webkitTransitionTimingFunction is 'ease-in'
PASS computedStyle.webkitTransitionTimingFunction is 'ease-in'
PASS style.transitionTimingFunction is 'ease-out'
PASS computedStyle.transitionTimingFunction is 'ease-out'
PASS style.webkitTransitionTimingFunction is 'ease-out'
PASS computedStyle.webkitTransitionTimingFunction is 'ease-out'
PASS style.transitionTimingFunction is 'ease-in-out'
PASS computedStyle.transitionTimingFunction is 'ease-in-out'
PASS style.webkitTransitionTimingFunction is 'ease-in-out'
PASS computedStyle.webkitTransitionTimingFunction is 'ease-in-out'
PASS style.transitionTimingFunction is 'cubic-bezier(0.25, 0.1, 0.25, 1)'
PASS computedStyle.transitionTimingFunction is 'cubic-bezier(0.25, 0.1, 0.25, 1)'
PASS style.webkitTransitionTimingFunction is 'cubic-bezier(0.25, 0.1, 0.25, 1)'
PASS computedStyle.webkitTransitionTimingFunction is 'cubic-bezier(0.25, 0.1, 0.25, 1)'
PASS style.transitionTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS computedStyle.transitionTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS style.webkitTransitionTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS computedStyle.webkitTransitionTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS style.transitionTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 3)'
PASS computedStyle.transitionTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 3)'
PASS style.webkitTransitionTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 3)'
PASS computedStyle.webkitTransitionTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 3)'
PASS style.transitionTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 3)'
PASS computedStyle.transitionTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 3)'
PASS style.webkitTransitionTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 3)'
PASS computedStyle.webkitTransitionTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 3)'
PASS style.transitionTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 0)'
PASS computedStyle.transitionTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 0)'
PASS style.webkitTransitionTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 0)'
PASS computedStyle.webkitTransitionTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 0)'
PASS style.transitionTimingFunction is 'cubic-bezier(0.2, 2, 0.2, -0.4)'
PASS computedStyle.transitionTimingFunction is 'cubic-bezier(0.2, 2, 0.2, -0.4)'
PASS style.webkitTransitionTimingFunction is 'cubic-bezier(0.2, 2, 0.2, -0.4)'
PASS computedStyle.webkitTransitionTimingFunction is 'cubic-bezier(0.2, 2, 0.2, -0.4)'
PASS style.transitionTimingFunction is 'cubic-bezier(0.2, -2, 0.2, 0.4)'
PASS computedStyle.transitionTimingFunction is 'cubic-bezier(0.2, -2, 0.2, 0.4)'
PASS style.webkitTransitionTimingFunction is 'cubic-bezier(0.2, -2, 0.2, 0.4)'
PASS computedStyle.webkitTransitionTimingFunction is 'cubic-bezier(0.2, -2, 0.2, 0.4)'
PASS style.transitionTimingFunction is 'step-start'
PASS computedStyle.transitionTimingFunction is 'steps(1, start)'
PASS style.webkitTransitionTimingFunction is 'step-start'
PASS computedStyle.webkitTransitionTimingFunction is 'steps(1, start)'
PASS style.transitionTimingFunction is 'step-end'
PASS computedStyle.transitionTimingFunction is 'steps(1, end)'
PASS style.webkitTransitionTimingFunction is 'step-end'
PASS computedStyle.webkitTransitionTimingFunction is 'steps(1, end)'
PASS style.transitionTimingFunction is 'steps(3, end)'
PASS computedStyle.transitionTimingFunction is 'steps(3, end)'
PASS style.webkitTransitionTimingFunction is 'steps(3, end)'
PASS computedStyle.webkitTransitionTimingFunction is 'steps(3, end)'
PASS style.transitionTimingFunction is 'steps(4, end)'
PASS computedStyle.transitionTimingFunction is 'steps(4, end)'
PASS style.webkitTransitionTimingFunction is 'steps(4, end)'
PASS computedStyle.webkitTransitionTimingFunction is 'steps(4, end)'
PASS style.transitionTimingFunction is 'steps(5, start)'
PASS computedStyle.transitionTimingFunction is 'steps(5, start)'
PASS style.webkitTransitionTimingFunction is 'steps(5, start)'
PASS computedStyle.webkitTransitionTimingFunction is 'steps(5, start)'
PASS style.transitionTimingFunction is 'ease-in-out, ease-in'
PASS computedStyle.transitionTimingFunction is 'ease-in-out, ease-in'
PASS style.webkitTransitionTimingFunction is 'ease-in-out, ease-in'
PASS computedStyle.webkitTransitionTimingFunction is 'ease-in-out, ease-in'
PASS style.transitionTimingFunction is 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS computedStyle.transitionTimingFunction is 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS style.webkitTransitionTimingFunction is 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS computedStyle.webkitTransitionTimingFunction is 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS style.transitionTimingFunction is 'steps(3, start), ease-in-out'
PASS computedStyle.transitionTimingFunction is 'steps(3, start), ease-in-out'
PASS style.webkitTransitionTimingFunction is 'steps(3, start), ease-in-out'
PASS computedStyle.webkitTransitionTimingFunction is 'steps(3, start), ease-in-out'
Invalid transition-timing-function values.
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS style.transitionTimingFunction is ''
PASS computedStyle.transitionTimingFunction is 'ease'
PASS style.webkitTransitionTimingFunction is ''
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
Valid transition-delay values.
PASS computedStyle.transitionDelay is '0s'
PASS computedStyle.webkitTransitionDelay is '0s'
PASS style.transitionDelay is '0s'
PASS computedStyle.transitionDelay is '0s'
PASS style.webkitTransitionDelay is '0s'
PASS computedStyle.webkitTransitionDelay is '0s'
PASS style.transitionDelay is '5s'
PASS computedStyle.transitionDelay is '5s'
PASS style.webkitTransitionDelay is '5s'
PASS computedStyle.webkitTransitionDelay is '5s'
PASS style.transitionDelay is '10ms'
PASS computedStyle.transitionDelay is '0.01s'
PASS style.webkitTransitionDelay is '10ms'
PASS computedStyle.webkitTransitionDelay is '0.01s'
PASS style.transitionDelay is '-10ms'
PASS computedStyle.transitionDelay is '-0.01s'
PASS style.webkitTransitionDelay is '-10ms'
PASS computedStyle.webkitTransitionDelay is '-0.01s'
PASS style.transitionDelay is '-10ms, 20s'
PASS computedStyle.transitionDelay is '-0.01s, 20s'
PASS style.webkitTransitionDelay is '-10ms, 20s'
PASS computedStyle.webkitTransitionDelay is '-0.01s, 20s'
Invalid transition-duration values.
PASS style.transitionDelay is ''
PASS computedStyle.transitionDelay is '0s'
PASS style.webkitTransitionDelay is ''
PASS computedStyle.webkitTransitionDelay is '0s'
PASS style.transitionDelay is ''
PASS computedStyle.transitionDelay is '0s'
PASS style.webkitTransitionDelay is ''
PASS computedStyle.webkitTransitionDelay is '0s'
PASS style.transitionDelay is ''
PASS computedStyle.transitionDelay is '0s'
PASS style.webkitTransitionDelay is ''
PASS computedStyle.webkitTransitionDelay is '0s'
PASS style.transitionDelay is ''
PASS computedStyle.transitionDelay is '0s'
PASS style.webkitTransitionDelay is ''
PASS computedStyle.webkitTransitionDelay is '0s'
PASS style.transitionDelay is ''
PASS computedStyle.transitionDelay is '0s'
PASS style.webkitTransitionDelay is ''
PASS computedStyle.webkitTransitionDelay is '0s'
PASS style.transitionDelay is ''
PASS computedStyle.transitionDelay is '0s'
PASS style.webkitTransitionDelay is ''
PASS computedStyle.webkitTransitionDelay is '0s'
Valid transition shorthand values.
PASS computedStyle.transition is 'all 0s ease 0s'
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is 'none'
PASS computedStyle.transition is 'none 0s ease 0s'
PASS style.webkitTransition is 'none'
PASS computedStyle.webkitTransition is 'none 0s ease 0s'
PASS checkTransitionShorthandValue() is true
PASS style.transition is 'none 20s'
PASS computedStyle.transition is 'none 20s ease 0s'
PASS style.webkitTransition is 'none 20s'
PASS computedStyle.webkitTransition is 'none 20s ease 0s'
PASS checkTransitionShorthandValue() is true
PASS style.transition is 'none 20s ease-in'
PASS computedStyle.transition is 'none 20s ease-in 0s'
PASS style.webkitTransition is 'none 20s ease-in'
PASS computedStyle.webkitTransition is 'none 20s ease-in 0s'
PASS checkTransitionShorthandValue() is true
PASS style.transition is 'opacity 20s ease-in'
PASS computedStyle.transition is 'opacity 20s ease-in 0s'
PASS style.webkitTransition is 'opacity 20s ease-in'
PASS computedStyle.webkitTransition is 'opacity 20s ease-in 0s'
PASS checkTransitionShorthandValue() is true
PASS style.transition is 'opacity 20s ease-in 10s'
PASS computedStyle.transition is 'opacity 20s ease-in 10s'
PASS style.webkitTransition is 'opacity 20s ease-in 10s'
PASS computedStyle.webkitTransition is 'opacity 20s ease-in 10s'
PASS computedStyle.transitionDuration is '20s'
PASS computedStyle.webkitTransitionDuration is '20s'
PASS computedStyle.transitionDelay is '10s'
PASS computedStyle.webkitTransitionDelay is '10s'
PASS checkTransitionShorthandValue() is true
PASS style.transition is 'opacity 20s ease-in 10s'
PASS computedStyle.transition is 'opacity 20s ease-in 10s'
PASS style.webkitTransition is 'opacity 20s ease-in 10s'
PASS computedStyle.webkitTransition is 'opacity 20s ease-in 10s'
PASS computedStyle.transitionDuration is '20s'
PASS computedStyle.webkitTransitionDuration is '20s'
PASS computedStyle.transitionDelay is '10s'
PASS computedStyle.webkitTransitionDelay is '10s'
PASS checkTransitionShorthandValue() is true
PASS style.transition is 'opacity 20s ease-in 10s, width 10s cubic-bezier(0.32, 0, 1, 1) 20s'
PASS computedStyle.transition is 'opacity 20s ease-in 10s, width 10s cubic-bezier(0.32, 0, 1, 1) 20s'
PASS style.webkitTransition is 'opacity 20s ease-in 10s, width 10s cubic-bezier(0.32, 0, 1, 1) 20s'
PASS computedStyle.webkitTransition is 'opacity 20s ease-in 10s, width 10s cubic-bezier(0.32, 0, 1, 1) 20s'
PASS computedStyle.transitionDuration is '20s, 10s'
PASS computedStyle.webkitTransitionDuration is '20s, 10s'
PASS computedStyle.transitionDelay is '10s, 20s'
PASS computedStyle.webkitTransitionDelay is '10s, 20s'
PASS checkTransitionShorthandValue() is true
PASS style.transition is 'opacity 20s ease-in 10s, all 20s ease-out'
PASS computedStyle.transition is 'opacity 20s ease-in 10s, all 20s ease-out 0s'
PASS style.webkitTransition is 'opacity 20s ease-in 10s, all 20s ease-out'
PASS computedStyle.webkitTransition is 'opacity 20s ease-in 10s, all 20s ease-out 0s'
PASS checkTransitionShorthandValue() is true
PASS style.transition is 'all 20s ease-out, opacity 20s ease-in 10s'
PASS computedStyle.transition is 'all 20s ease-out 0s, opacity 20s ease-in 10s'
PASS style.webkitTransition is 'all 20s ease-out, opacity 20s ease-in 10s'
PASS computedStyle.webkitTransition is 'all 20s ease-out 0s, opacity 20s ease-in 10s'
PASS checkTransitionShorthandValue() is true
Invalid transition shorthand values.
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS style.transition is ''
PASS computedStyle.transition is 'all 0s ease 0s'
PASS style.webkitTransition is ''
PASS computedStyle.webkitTransition is 'all 0s ease 0s'
PASS successfullyParsed is true

TEST COMPLETE
